Introduce libtiny_ui
