+++ libtiny_ui/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "libtiny_ui"
+version = "0.1.0"
+authors = ["Ömer Sinan Ağacan <omeragacan@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
